Adjust styling and improve scalability
The changes made adjust some minor text and margin styles within the chart-calendar.html. Additionally, the aspectRatio attribute was included in the setDimensions function. This helps in more dynamic rendering on various screen sizes. The 'drawLegend' function now also takes 'aspectRatio' as a parameter, in order to adapt the legend properly to different screens.