Add region selection dropdown and data visualisations
Added a region selection dropdown to the histogram chart. Also populated this dropdown with unique regions from the parsed data. Improvements to the chart setup functions and data population have been made, along with additions of histogram, timeseries, and map charts. The status color for the 'Map' section in the 'index.html' file now shows as red instead of orange to reflect its 'not started' status.