Update HTML formatting and chart hotspots with tooltips
HTML formatting across files was adjusted, and whitespace has been added. The relative positions of the elements in the histogram chart were updated to give clearer and more accurate data representation. The chart's hotspot now features tooltips to provide users with the name of the local area when hovering over the chart. The completion status of the histogram chart has now been updated to "complete".